==> analysis/integration/mc_3c/02a.prepare_mc.ipynb <==
## Import

from wmb import cemba
from ALLCools.mcds import MCDS
import pandas as pd
from ALLCools.clustering import log_scale

var_dim = 'chrom100k'
cluster_col = 'L1'
chrom_to_remove = ['chrX', 'chrY', 'chrM', 'chrL']
downsample = 1000
mc_type = 'CHN'

## Select cells

cells = pd.read_csv('mc_cells.txt', index_col=0, header=None).index
cells.name = 'cell'
cells.size

## Get adata with CEF

mcds = MCDS.open(cemba.CEMBA_SNMC_MCDS_PATH, var_dim=var_dim, use_obs=cells)

mcds

cef = pd.read_csv('CEF.csv', header=None, index_col=0).squeeze()
cef = cef[cef].index
cef.name = var_dim

mcds.add_feature_selection_column(cef, f'{var_dim}_cef')
adata = mcds.get_adata(mc_type=mc_type,
                       select_hvf=f'{var_dim}_cef',
                       split_large_chunks=False)

annot = cemba.get_mc_annot()
adata.obs[cluster_col] = annot[cluster_col].to_pandas()

adata

# log mC fraction and scale features
log_scale(adata, with_mean=True)

adata.write_h5ad('mc_input.h5ad')
